requirements.txt
numpy
obs-websocket-py
sounddevice
